Add files via upload
Computer Animation Project where I designed a WASP and implemented a 3D version of it in UNITY by precisely inputting vertices to draw it. It should walk normally... unfortunately I seem to have misplaced the final version and it sort of wiggles in this version. Maybe I will fix it later haha